hopefully handle resetting shipping on load
